fix: 선택된 체크박스 핸들하는 state 공통부모로 올림
