+++ src/components/ticketList/DropDown/FlightDropDown.tsx
@@ -9,25 +9,23 @@ import { List, Path, Price, StyledInput, StyledLabel, StyledText } from './style
 
 // API 연결 필요
 const FlightDropDown = (props: DropDownProps) => {
-  const { isOpen, onToggle, flightList } = props;
-  const [selectedAirName, setSelectedAirName] = useState<string[]>([]);
+  const { isOpen, onToggle, flightList, checkboxHandler } = props;
+  // const [selectedAirName, setSelectedAirName] = useState<string[]>([]);
 
-  const onClickCheckbox = (e: React.MouseEvent<HTMLInputElement>) => {
-    const inputElement = e.target as HTMLInputElement;
-    const selectedName = inputElement.name;
+  // const onClickCheckbox = (e: React.MouseEvent<HTMLInputElement>) => {
+  //   const inputElement = e.target as HTMLInputElement;
+  //   const selectedName = inputElement.name;
 
-    setSelectedAirName((prev) => {
-      const isSelected = prev.includes(selectedName);
+  //   setSelectedAirName((prev) => {
+  //     const isSelected = prev.includes(selectedName);
 
-      if (isSelected) {
-        return prev.filter((name) => name !== selectedName);
-      } else {
-        return [...prev, selectedName];
-      }
-    });
-  };
-
-  console.log(selectedAirName);
+  //     if (isSelected) {
+  //       return prev.filter((name) => name !== selectedName);
+  //     } else {
+  //       return [...prev, selectedName];
+  //     }
+  //   });
+  // };
 
   return (
     <DropDownContainer>
@@ -40,7 +38,7 @@ const FlightDropDown = (props: DropDownProps) => {
         <List>
           {flightList?.map((item) => (
             <StyledLabel htmlFor={item.airId.toString()} key={item.airId}>
-              <StyledInput type="checkbox" id={item.airId.toString()} name={item.airName} onClick={onClickCheckbox} />
+              <StyledInput type="checkbox" id={item.airId.toString()} name={item.airName} onClick={checkboxHandler} />
               <StyledText>
                 <Path>{item.airName}</Path>
                 <Price>{item.minPriceString}</Price>
