+++ src/pages/TicketList.tsx
@@ -1,3 +1,5 @@
+import { useState } from 'react';
+
 import SearchBar from '../components/common/SearchBar';
 import AdvImg from '../components/ticketList/AdvImg';
 import BackgroundImg from '../components/ticketList/BackgroundImg';
@@ -7,13 +9,30 @@ import Nav from '../components/ticketList/Nav';
 import Ticket from '../components/ticketList/Ticket';
 
 const TicketList = () => {
+  const [selectedAirName, setSelectedAirName] = useState<string[]>([]);
+
+  const onClickCheckbox = (e: React.MouseEvent<HTMLInputElement>) => {
+    const inputElement = e.target as HTMLInputElement;
+    const selectedName = inputElement.name;
+
+    setSelectedAirName((prev) => {
+      const isSelected = prev.includes(selectedName);
+
+      if (isSelected) {
+        return prev.filter((name) => name !== selectedName);
+      } else {
+        return [...prev, selectedName];
+      }
+    });
+  };
+
   return (
     <>
       <Nav />
       <BackgroundImg />
       <SearchBar />
       <AdvImg />
-      <DropDown />
+      <DropDown checkboxHandler={onClickCheckbox}/>
       <Ticket rangeStart={0} rangeEnd={4} />
       <HotelCarousel />
       <Ticket rangeStart={4} />
